moved:wirechat widget from to components folder
