+++ src/Livewire/Components/NewChat.php
@@ -4,7 +4,7 @@ namespace Namu\WireChat\Livewire\Components;
 
 use Namu\WireChat\Facades\WireChat;
 use Namu\WireChat\Livewire\Modals\ModalComponent;
-use Namu\WireChat\Livewire\Widgets\WireChat as WidgetsWireChat;
+use Namu\WireChat\Livewire\Components\WireChat as WidgetsWireChat;
 use Namu\WireChat\Traits\Widget;
 
 class NewChat extends ModalComponent
@@ -42,6 +42,7 @@ class NewChat extends ModalComponent
         }
     }
 
+    
     public function createConversation($id, string $class)
     {
